campus/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
djangorestframework==3.14.0
mysqlclient==2.1.1
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
